fix vc2010 warnings

+++ src/gui/GuiTextEntry.cpp
@@ -130,7 +130,7 @@ bool TextEntry::OnKeyPress(const SDL_keysym *sym)
 void TextEntry::GetSizeRequested(float size[2])
 {
 	// XXX this 1.5f should be PARAGRAPH_SPACING (currently #define'd in TextureFont.h)
-	size[1] = (m_newlineCount*1.5f+1.0f)*Gui::Screen::GetFontHeight(m_font) + 2.0;
+	size[1] = (m_newlineCount*1.5f+1.0f)*Gui::Screen::GetFontHeight(m_font) + 2.0f;
 }
 
 bool TextEntry::OnMouseDown(MouseButtonEvent *e)
@@ -178,9 +178,9 @@ void TextEntry::Draw()
 
 	glColor3f(1,0,0);
 	if (curs_x - m_scroll > size[0]*0.75f) {
-		m_scroll += size[0]*0.25f;
+		m_scroll += int(size[0]*0.25f);
 	} else if (curs_x - m_scroll < size[0]*0.25f) {
-		m_scroll -= size[0]*0.25f;
+		m_scroll -= int(size[0]*0.25f);
 		if (m_scroll < 0) m_scroll = 0;
 	}
 
@@ -205,10 +205,10 @@ void TextEntry::Draw()
 	Gui::Screen::RenderString(m_text, 1.0f - m_scroll, 1.0f, m_font);
 
 	/* Cursor */
-	glColor3f(0.5,0.5,0.5);
+	glColor3f(0.5f,0.5f,0.5f);
 	glBegin(GL_LINES);
-		glVertex2f(curs_x + 1.0f - m_scroll, curs_y - Gui::Screen::GetFontHeight(m_font) - 1.0);
-		glVertex2f(curs_x + 1.0f - m_scroll, curs_y + 1.0);
+		glVertex2f(curs_x + 1.0f - m_scroll, curs_y - Gui::Screen::GetFontHeight(m_font) - 1.0f);
+		glVertex2f(curs_x + 1.0f - m_scroll, curs_y + 1.0f);
 	glEnd();
 	
 	EndClipping();
